requirements.txt
absl-py==1.4.0
actionlib==1.14.0
addict==2.4.0
aiohttp==3.8.4
aiosignal==1.3.1
aliyun-python-sdk-core==2.15.1
aliyun-python-sdk-kms==2.16.3
angles==1.9.13
async-timeout==4.0.2
attrs==22.2.0
audioread==3.0.1
av==10.0.0
blessed==1.20.0
bondpy==1.8.6
cachetools==5.3.0
camera-calibration==1.17.0
camera-calibration-parsers==1.12.0
catkin==0.8.10
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi==1.16.0
charset-normalizer==3.1.0
chumpy==0.70
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
controller-manager==0.19.6
controller-manager-msgs==0.19.6
crcmod==1.7
cryptography==43.0.0
cv-bridge==1.16.2
cycler==0.12.1
Cython==3.0.10
decorator==4.4.2
decord==0.6.0
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
dynamic-reconfigure==1.7.3
easydict==1.13
einops==0.6.0
filelock==3.14.0
fire==0.6.0
fonttools==4.53.1
frozenlist==1.3.3
fsspec==2023.3.0
fvcore==0.1.5.post20221221
gazebo_plugins==2.9.2
gazebo_ros==2.9.2
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
google-auth==2.16.3
google-auth-oauthlib==0.4.6
gpustat==1.1.1
grpcio==1.51.3
idna==3.4
image-geometry==1.16.2
imageio==2.34.2
imageio-ffmpeg==0.5.1
importlib_metadata==8.2.0
importlib_resources==6.4.0
interactive-markers==1.12.0
iopath==0.1.10
jmespath==0.10.0
joblib==1.4.2
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
json-tricks==3.17.3
kiwisolver==1.4.5
laser_geometry==1.6.7
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.9.1
mdurl==0.1.2
message-filters==1.16.0
mmaction2==0.24.1
mmcv==1.6.0
mmengine==0.10.4
model-index==0.1.11
moviepy==1.0.3
msgpack==1.0.8
multidict==6.0.4
munkres==1.1.4
networkx==3.2.1
numba==0.60.0
numpy==1.24.2
nvidia-ml-py==12.555.43
oauthlib==3.2.2
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.1
ordered-set==4.1.0
oss2==2.17.0
packaging==24.1
pandas==1.5.3
Pillow==9.4.0
platformdirs==4.2.2
pooch==1.8.2
portalocker==2.7.0
proglog==0.1.10
protobuf==4.22.1
psutil==6.0.0
ptflops==0.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.8
pycparser==2.22
pycryptodome==3.20.0
pyDeprecate==0.3.2
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.8.2
python-qt-binding==0.4.4
python-speech-features==0.6
pytorch-lightning==1.7.7
pytz==2023.4
pyworld==0.3.5
PyYAML==6.0
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
requests==2.28.2
requests-oauthlib==1.3.1
resource_retriever==1.12.7
rich==13.4.2
rosbag==1.16.0
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosgraph==1.16.0
roslaunch==1.16.0
roslib==1.15.8
roslint==0.12.0
roslz4==1.16.0
rosmake==1.15.8
rosmaster==1.16.0
rosmsg==1.16.0
rosnode==1.16.0
rosparam==1.16.0
rospy==1.16.0
rosservice==1.16.0
rostest==1.16.0
rostopic==1.16.0
rosunit==1.15.8
roswtf==1.16.0
rqt-image-view==0.4.17
rqt-moveit==0.5.10
rqt-reconfigure==0.5.5
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.14
rqt-rviz==0.7.0
rqt_action==0.4.9
rqt_bag==0.5.1
rqt_bag_plugins==0.5.1
rqt_console==0.4.11
rqt_dep==0.4.12
rqt_graph==0.4.14
rqt_gui==0.5.3
rqt_gui_py==0.5.3
rqt_launch==0.4.9
rqt_logger_level==0.4.11
rqt_msg==0.4.10
rqt_nav_view==0.5.7
rqt_plot==0.4.13
rqt_pose_view==0.5.11
rqt_publisher==0.4.10
rqt_py_common==0.5.3
rqt_py_console==0.4.10
rqt_robot_steering==0.5.12
rqt_runtime_monitor==0.5.9
rqt_service_caller==0.4.10
rqt_shell==0.4.11
rqt_srv==0.4.9
rqt_tf_tree==0.6.3
rqt_top==0.4.10
rqt_topic==0.4.13
rqt_web==0.4.10
rsa==4.9
rviz==1.14.20
scikit-image==0.24.0
scikit-learn==1.5.0
scipy==1.13.1
sensor-msgs==1.13.1
shapely==2.0.5
six==1.16.0
smach==2.5.2
smach-ros==2.5.2
smclib==1.8.6
soundfile==0.12.1
soxr==0.4.0
tabulate==0.9.0
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
termcolor==2.3.0
terminaltables==3.1.10
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.6
tf2-kdl==0.7.6
tf2-py==0.7.6
tf2-ros==0.7.6
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tifffile==2024.8.30
toml==0.10.2
tomli==2.0.1
topic-tools==1.16.0
torch==1.13.1+cu117
torchaudio==0.13.1+cu117
torchmetrics==0.11.4
torchvision==0.14.1+cu117
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.13
Werkzeug==2.2.3
xacro==1.14.16
xtcocotools==1.14.3
yacs==0.1.8
yapf==0.40.2
yarl==1.8.2
zipp==3.15.0
